choropleth legend
