+++ components/indicator-visualizations/choropleth/legend/choropleth/component.tsx
