+++ components/indicator-visualizations/choropleth/legend/choropleth/index.ts
