+++ components/indicator-visualizations/choropleth/legend/choropleth/types.d.ts
