+++ components/indicator-visualizations/choropleth/legend/gradient/component.tsx
