+++ components/indicator-visualizations/choropleth/legend/gradient/index.ts
