+++ components/indicator-visualizations/choropleth/legend/gradient/types.d.ts
