+++ components/indicator-visualizations/choropleth/legend/item/component.tsx
