+++ components/indicator-visualizations/choropleth/legend/item/index.ts
