+++ components/indicator-visualizations/choropleth/legend/mock.tsx
