+++ components/indicator-visualizations/choropleth/legend/sortable/list/component.tsx
