+++ components/indicator-visualizations/choropleth/legend/sortable/list/index.ts
