+++ components/indicator-visualizations/choropleth/legend/sortable/list/item/component.tsx
