+++ components/indicator-visualizations/choropleth/legend/sortable/list/item/index.ts
