--- components/indicator-visualizations/choropleth/legend2/sortable/list/item/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
