+++ components/legend/component.tsx
@@ -3,15 +3,15 @@ import cx from 'classnames';
 
 import { useColors } from 'hooks/utils';
 
-interface LegendProps {
+interface WidgetLegendProps {
   categories: string[],
   className?: string,
 }
 
-const Legend: FC<LegendProps> = ({
+const Legend: FC<WidgetLegendProps> = ({
   categories = [],
   className = '',
-}: LegendProps) => {
+}: WidgetLegendProps) => {
   const colors = useColors(categories.length);
   return (
     <div className={cx('inline-flex flex-col justify-start text-center bg-white rounded-md border-gray5 border-6 hover:opacity-90 px-1.5 text-gray1 max-h-64',
